+++ lib/td_dd/executions/audit.ex
@@ -10,11 +10,15 @@ defmodule TdDd.Executions.Audit do
   @doc """
   Publishes an `:execution_group_created` event. Should be called using `Ecto.Multi.run/5`.
   """
-  def execution_group_created(_repo, %{profile_group: group}, %{changes: changes} = _changeset) do
-    execution_group_created(group, changes)
+  def execution_group_created(_repo, %{profile_group: %{created_by_id: created_by_id} = group}) do
+    handle_execution_group_created(group, %{created_by_id: created_by_id})
   end
 
-  defp execution_group_created(%{id: id, executions: executions} = _group, %{} = changes) do
+  def execution_group_created(_repo, %{profile_group: group}, %{changes: changes} = _changeset) do
+    handle_execution_group_created(group, changes)
+  end
+
+  defp handle_execution_group_created(%{id: id, executions: executions} = _group, %{} = changes) do
     user_id = Map.get(changes, :created_by_id, 0)
 
     executions = Enum.map(executions, &Map.take(&1, [:id, :data_structure_id]))
@@ -22,5 +26,5 @@ defmodule TdDd.Executions.Audit do
     publish("execution_group_created", "profile_execution_group", id, user_id, payload)
   end
 
-  defp execution_group_created(_, _), do: {:error, :invalid}
+  defp handle_execution_group_created(_, _), do: {:error, :invalid}
 end