+++ test/td_dd/data_structures/bulk_update_test.exs
@@ -86,14 +86,6 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
           "widget" => "enriched_text"
         },
         %{
-          "cardinality" => "*",
-          "label" => "Urls One Or None",
-          "name" => "urls_one_or_none",
-          "type" => "url",
-          "values" => nil,
-          "widget" => "pair_list"
-        },
-        %{
           "cardinality" => "?",
           "label" => "Numeric",
           "name" => "integer",
@@ -114,6 +106,38 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
           "type" => "hierarchy",
           "values" => %{"hierarchy" => %{"id" => 1}},
           "widget" => "dropdown"
+        },
+        %{
+          "name" => "father",
+          "type" => "string",
+          "label" => "father",
+          "values" => %{"fixed" => ["a1", "a2", "b1", "b2"]},
+          "widget" => "dropdown",
+          "default" => %{"value" => "", "origin" => "default"},
+          "cardinality" => "?",
+          "subscribable" => false,
+          "ai_suggestion" => false
+        },
+        %{
+          "name" => "son",
+          "type" => "string",
+          "label" => "son",
+          "values" => %{
+            "switch" => %{
+              "on" => "father",
+              "values" => %{
+                "a1" => ["a11", "a12", "a13"],
+                "a2" => ["a21", "a22", "a23"],
+                "b1" => ["b11", "b12", "b13"],
+                "b2" => ["b21", "b22", "b23"]
+              }
+            }
+          },
+          "widget" => "dropdown",
+          "default" => %{"value" => "", "origin" => "default"},
+          "cardinality" => "?",
+          "subscribable" => false,
+          "ai_suggestion" => false
         }
       ]
     }
@@ -154,6 +178,29 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
           "type" => "string",
           "values" => %{"fixed" => ["pear", "banana", "apple", "peach"]},
           "widget" => "checkbox"
+        }
+      ]
+    }
+  ]
+
+  @url_template [
+    %{
+      "name" => "url_template",
+      "fields" => [
+        %{
+          "cardinality" => "?",
+          "label" => "Text",
+          "name" => "text",
+          "type" => "string",
+          "widget" => "string"
+        },
+        %{
+          "cardinality" => "*",
+          "label" => "Urls None or More",
+          "name" => "urls_none_or_more",
+          "type" => "url",
+          "values" => nil,
+          "widget" => "pair_list"
         }
       ]
     }
@@ -632,7 +679,7 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
                |> BulkUpdate.do_csv_bulk_update(user_id)
 
       ids = Map.keys(update_notes)
-      assert length(ids) == 10
+      assert length(ids) == 14
       assert Enum.all?(ids, fn id -> id in structure_ids end)
 
       assert %{
@@ -673,16 +720,13 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
                get_df_content_from_ext_id("ex_id6")
 
       text = to_enriched_text("Enriched text")
-      url = to_content_url("https://www.google.es")
 
       assert %{
                "enriched_text" => %{"value" => ^text, "origin" => "file"},
-               "urls_one_or_none" => %{"value" => ^url, "origin" => "file"},
                "integer" => %{"value" => 3, "origin" => "file"}
              } = get_df_content_from_ext_id("ex_id7")
 
       assert %{
-               "urls_one_or_none" => %{"value" => ^url, "origin" => "file"},
                "integer" => %{"value" => 2, "origin" => "file"}
              } = get_df_content_from_ext_id("ex_id8")
 
@@ -697,9 +741,57 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
 
       assert %{
                "hierarchy_name_1" => %{"value" => ^key_node_2, "origin" => "file"},
-               "hierarchy_name_2" => %{"value" => [^key_node_2, ^key_node_1], "origin" => "file"},
-               "urls_one_or_none" => %{"value" => _, "origin" => "file"}
+               "hierarchy_name_2" => %{"value" => [^key_node_2, ^key_node_1], "origin" => "file"}
              } = get_df_content_from_ext_id("ex_id10")
+
+      assert %{
+               "text" => %{"value" => "URL Single url without name"},
+               "urls_none_or_more" => %{
+                 "value" => [
+                   %{
+                     "url_name" => "",
+                     "url_value" => "https://www.google.es"
+                   }
+                 ],
+                 "origin" => "file"
+               }
+             } = get_df_content_from_ext_id("ex_id16")
+
+      assert %{
+               "text" => %{"value" => "URL Single url with name"},
+               "urls_none_or_more" => %{
+                 "value" => [
+                   %{"url_name" => "Google", "url_value" => "https://www.google.es"}
+                 ],
+                 "origin" => "file"
+               }
+             } = get_df_content_from_ext_id("ex_id17")
+
+      assert %{
+               "text" => %{"value" => "URL Multiple urls"},
+               "urls_none_or_more" => %{
+                 "value" => [
+                   %{"url_name" => "Google", "url_value" => "https://www.google.es"},
+                   %{
+                     "url_name" => "",
+                     "url_value" => "https://www.google.es"
+                   },
+                   %{
+                     "url_name" => "",
+                     "url_value" => "https://www.google.es"
+                   }
+                 ],
+                 "origin" => "file"
+               }
+             } = get_df_content_from_ext_id("ex_id18")
+
+      assert %{
+               "text" => %{"value" => "URL No url"},
+               "urls_none_or_more" => %{
+                 "value" => [%{"url_name" => "", "url_value" => ""}],
+                 "origin" => "file"
+               }
+             } = get_df_content_from_ext_id("ex_id19")
 
       assert [{:reindex, :structures, ^ids}] = IndexWorkerMock.calls()
     end
@@ -838,7 +930,8 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
 
       [_, errored_notes] = BulkUpdate.split_succeeded_errors(update_notes)
 
-      [first_errored_note | _] = Enum.map(errored_notes, fn {_k, v} -> v end)
+      [first_errored_note, second_errored_note | _] =
+        Enum.map(errored_notes, fn {_k, v} -> v end)
 
       assert {:error,
               {%{
@@ -846,13 +939,193 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
                    df_content:
                      {_,
                       [
-                        hierarchy_name_1: {"has more than one node children_2"},
-                        hierarchy_name_2: {"has more than one node children_2"}
+                        hierarchy_name_2: {"hierarchy"},
+                        hierarchy_name_1: {"hierarchy"}
                       ]}
                  ]
-               }, _}} = first_errored_note
+               },
+               %TdDd.DataStructures.DataStructure{
+                 external_id: "ex_id9"
+               }}} = first_errored_note
 
-      assert [{:reindex, :structures, [_]}] = IndexWorkerMock.calls()
+      assert {:error,
+              {%{
+                 errors: [
+                   df_content:
+                     {_,
+                      [
+                        hierarchy_name_2: {"has more than one node children_2"},
+                        hierarchy_name_1: {"has more than one node children_2"}
+                      ]}
+                 ]
+               },
+               %TdDd.DataStructures.DataStructure{
+                 external_id: "ex_id10"
+               }}} = second_errored_note
+
+      assert [{:reindex, :structures, [_ | _]}] = IndexWorkerMock.calls()
+    end
+
+    test "returns error on dependant invalid content" do
+      IndexWorkerMock.clear()
+
+      previus_ex_id9 =
+        get_df_content_from_ext_id("ex_id9")
+
+      previus_ex_id10 =
+        get_df_content_from_ext_id("ex_id10")
+
+      %{user_id: user_id} = build(:claims)
+      upload = %{path: "test/fixtures/upload_invalid_dependant.csv"}
+
+      {:ok, %{update_notes: update_notes}} =
+        upload
+        |> BulkUpdate.from_csv(@default_lang)
+        |> BulkUpdate.do_csv_bulk_update(user_id)
+
+      [_, errored_notes] =
+        BulkUpdate.split_succeeded_errors(update_notes)
+
+      [first_errored_note, second_errored_note | _] =
+        Enum.map(errored_notes, fn {_k, v} -> v end)
+
+      [first_errored_note_id, second_errored_note_id | _] =
+        Enum.map(errored_notes, fn {k, _v} -> k end)
+
+      assert {:error,
+              {%{
+                 errors: [
+                   df_content:
+                     {_,
+                      [
+                        son: {"is invalid", [validation: :inclusion, enum: ["b11", "b12", "b13"]]}
+                      ]}
+                 ]
+               },
+               %TdDd.DataStructures.DataStructure{
+                 external_id: "ex_id9"
+               }}} = first_errored_note
+
+      assert {:error,
+              {%{
+                 errors: [
+                   df_content:
+                     {_,
+                      [
+                        son: {"is invalid", [validation: :inclusion, enum: ["a11", "a12", "a13"]]}
+                      ]}
+                 ]
+               },
+               %TdDd.DataStructures.DataStructure{
+                 external_id: "ex_id10"
+               }}} = second_errored_note
+
+      assert previus_ex_id9 ==
+               get_df_content_from_ext_id("ex_id9")
+
+      assert previus_ex_id10 ==
+               get_df_content_from_ext_id("ex_id10")
+
+      assert [{:reindex, :structures, [first_errored_note_id, second_errored_note_id]}] ==
+               IndexWorkerMock.calls()
+    end
+
+    test "create valid notes and update invalid notes using csv_bulk_update" do
+      IndexWorkerMock.clear()
+
+      previus_ex_id9 = get_df_content_from_ext_id("ex_id9")
+
+      previus_ex_id10 = get_df_content_from_ext_id("ex_id10")
+
+      %{user_id: user_id} = build(:claims)
+      upload = %{path: "test/fixtures/upload_valid_dependant.csv"}
+
+      {:ok, %{update_notes: update_notes}} =
+        upload
+        |> BulkUpdate.from_csv(@default_lang)
+        |> BulkUpdate.do_csv_bulk_update(user_id)
+
+      [_created_notes, errored_notes] =
+        BulkUpdate.split_succeeded_errors(update_notes)
+
+      assert map_size(errored_notes) == 0
+
+      intermediate_ex_id9 =
+        get_df_content_from_ext_id("ex_id9")
+
+      intermediate_ex_id10 =
+        get_df_content_from_ext_id("ex_id10")
+
+      assert previus_ex_id9 != intermediate_ex_id9 and
+               %{
+                 "father" => %{"origin" => "file", "value" => "b1"},
+                 "son" => %{"origin" => "file", "value" => "b11"}
+               } == intermediate_ex_id9
+
+      assert previus_ex_id10 != intermediate_ex_id10 and
+               %{
+                 "father" => %{"origin" => "file", "value" => "a1"},
+                 "son" => %{"origin" => "file", "value" => "a11"}
+               } == intermediate_ex_id10
+
+      assert [{:reindex, :structures, [_, _]}] = IndexWorkerMock.calls()
+
+      IndexWorkerMock.clear()
+
+      upload = %{path: "test/fixtures/upload_invalid_dependant.csv"}
+
+      {:ok, %{update_notes: update_notes}} =
+        upload
+        |> BulkUpdate.from_csv(@default_lang)
+        |> BulkUpdate.do_csv_bulk_update(user_id)
+
+      [updated_notes, errored_notes] =
+        BulkUpdate.split_succeeded_errors(update_notes)
+
+      assert map_size(updated_notes) == 0
+
+      [first_errored_note, second_errored_note | _] =
+        Enum.map(errored_notes, fn {_k, v} -> v end)
+
+      [first_errored_note_id, second_errored_note_id | _] =
+        Enum.map(errored_notes, fn {k, _v} -> k end)
+
+      assert {:error,
+              {%{
+                 errors: [
+                   df_content:
+                     {_,
+                      [
+                        son: {"is invalid", [validation: :inclusion, enum: ["b11", "b12", "b13"]]}
+                      ]}
+                 ]
+               },
+               %TdDd.DataStructures.DataStructure{
+                 external_id: "ex_id9"
+               }}} = first_errored_note
+
+      assert {:error,
+              {%{
+                 errors: [
+                   df_content:
+                     {_,
+                      [
+                        son: {"is invalid", [validation: :inclusion, enum: ["a11", "a12", "a13"]]}
+                      ]}
+                 ]
+               },
+               %TdDd.DataStructures.DataStructure{
+                 external_id: "ex_id10"
+               }}} = second_errored_note
+
+      assert intermediate_ex_id9 ==
+               get_df_content_from_ext_id("ex_id9")
+
+      assert intermediate_ex_id10 ==
+               get_df_content_from_ext_id("ex_id10")
+
+      assert [{:reindex, :structures, [^first_errored_note_id, ^second_errored_note_id]}] =
+               IndexWorkerMock.calls()
     end
 
     test "accept file utf8 with bom" do
@@ -1002,11 +1275,16 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
     %{id: id_t1, name: type1} = CacheHelpers.insert_template(content: @c1)
     %{id: id_t2, name: type2} = CacheHelpers.insert_template(content: @c2)
     %{id: id_t3, name: type3} = CacheHelpers.insert_template(content: @c3)
+
+    %{id: url_template_id, name: url_template_name} =
+      CacheHelpers.insert_template(content: @url_template)
+
     domain = CacheHelpers.insert_domain()
 
     insert(:data_structure_type, name: type1, template_id: id_t1)
     insert(:data_structure_type, name: type2, template_id: id_t2)
     insert(:data_structure_type, name: type3, template_id: id_t3)
+    insert(:data_structure_type, name: url_template_name, template_id: url_template_id)
 
     sts1 =
       Enum.map(1..5, fn id ->
@@ -1032,7 +1310,14 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
         valid_structure_note(type3, data_structure, opts)
       end)
 
-    [sts: sts1 ++ sts2 ++ sts3]
+    url_sts =
+      Enum.map(16..19, fn id ->
+        data_structure = insert(:data_structure, external_id: "ex_id#{id}")
+
+        valid_structure_note(url_template_name, data_structure, [])
+      end)
+
+    [sts: sts1 ++ sts2 ++ sts3 ++ url_sts]
   end
 
   defp insert_i18n_messages(_) do
@@ -1050,10 +1335,6 @@ defmodule TdDd.DataStructures.BulkUpdateTest do
       %{message_id: "fields.label_i18n_test.Checkbox Fixed.banana", definition: "plátano"},
       %{message_id: "fields.label_i18n_test.Checkbox Fixed.apple", definition: "manzana"}
     ])
-  end
-
-  defp to_content_url(url) do
-    [%{"url_name" => url, "url_value" => url}]
   end
 
   defp to_enriched_text(text) do